Split away the deprecated beta notice test
Move the BetaNoticeInserterTest into `deprecated_usage_test.rb`, and
include an assertion on the warning notice (which also silences it in
tests).
