+++ test/typical_usage_test.rb
@@ -253,24 +253,6 @@ module TypicalUsage
     end
   end
 
-  class BetaNoticeInserterTest < SlimmerIntegrationTest
-    def test_should_add_beta_warnings
-      given_response 200, %{
-        <html>
-          <body class="wibble">
-            <div id="wrapper">The body of the page</div>
-          </body>
-        </html>
-      }, {Slimmer::Headers::BETA_HEADER => '1'}
-
-      # beta notice after cookie bar
-      assert_rendered_in_template "body.beta.wibble #global-header + div.beta-notice"
-
-      # beta notice before footer
-      assert_rendered_in_template "body.beta.wibble div.beta-notice.js-footer + #footer"
-    end
-  end
-
   class BetaLabelInserterTest < SlimmerIntegrationTest
     def test_should_add_beta_label
       given_response 200, %{
